fix: move relation builder
